+++ src/main/java/xyz/falado/whs/model/WHSSubProjectExample.java
@@ -144,7 +144,7 @@ public class WHSSubProjectExample {
             return (Criteria) this;
         }
 
-        public Criteria andIdIn(List<Integer> values) {
+        public Criteria andIdIn(List<Long> values) {
             addCriterion("id in", values, "id");
             return (Criteria) this;
         }
